==> 01_IFresize HEcrop.ipynb <==
import glob
import math
import numpy as np
import os
import pandas as pd
from PIL import Image
import PIL
from skimage.io import imread,imshow,imsave 

os.chdir ('/Users/kim/Data/CRUK-OHSU_onedrive final 2022-09-30/Manuscript/IMC-IF method paper/Cell_report_method_resource/github')

print('library import done')

### IF resize

IFpaths=sorted(glob.glob('Data/IMC EMR (1)/ROI001*/2. IF/*.czi.tif'))

# CPU each file with specific folder 

IFpath=IFpaths[-1]
name=IFpath.split('/')[2]
print(name)

IForg=imread(IFpath)
DAPI=IForg[:,:,0]
Ecad=IForg[:,:,1]
CD45=IForg[:,:,2] 

width=math.ceil(DAPI.shape[0]*0.325)
height=math.ceil(DAPI.shape[1]*0.325)    

DAPI_re = Image.fromarray(DAPI)
DAPI_resize=DAPI_re.resize(size=(height,width),resample=PIL.Image.NEAREST)
DAPI_resize_array = np.array(DAPI_resize)

Ecad_re = Image.fromarray(Ecad)
Ecad_resize=Ecad_re.resize(size=(height,width),resample=PIL.Image.NEAREST)
Ecad_resize_array = np.array(Ecad_resize)

CD45_re = Image.fromarray(CD45)
CD45_resize=CD45_re.resize(size=(height,width),resample=PIL.Image.NEAREST)
CD45_resize_array = np.array(CD45_resize)

imsave(f'Data/IMC EMR (1)/{name}/resized_IF/{name}_DAPI.tif',DAPI_resize_array)  
print(DAPI_resize_array.shape)
imsave(f'Data/IMC EMR (1)/{name}/resized_IF/{name}_Ecad.tif',Ecad_resize_array) 
print(Ecad_resize_array.shape)
imsave(f'Data/IMC EMR (1)/{name}/resized_IF/{name}_CD45.tif',CD45_resize_array) 
print(CD45_resize_array.shape)

# GPU Loop 

n=0
for i in range(len(IFpaths)): 
    IFpath=IFpaths[n]
    name=IFpath.split('/')[2]
    IForg=imread(IFpath)
    DAPI=IForg[:,:,0]
    Ecad=IForg[:,:,1]
    CD45=IForg[:,:,2] 

    width=math.ceil(DAPI.shape[0]*0.325)
    height=math.ceil(DAPI.shape[1]*0.325)    

    DAPI_re = Image.fromarray(DAPI)
    DAPI_resize=DAPI_re.resize(size=(height,width),resample=PIL.Image.NEAREST)
    DAPI_resize_array = np.array(DAPI_resize)
    
    Ecad_re = Image.fromarray(Ecad)
    Ecad_resize=Ecad_re.resize(size=(height,width),resample=PIL.Image.NEAREST)
    Ecad_resize_array = np.array(Ecad_resize)
    
    CD45_re = Image.fromarray(CD45)
    CD45_resize=CD45_re.resize(size=(height,width),resample=PIL.Image.NEAREST)
    CD45_resize_array = np.array(CD45_resize)
    
    path=f'new_July_2021/raw_data/{name}/resized_IF/'
    createFolder(path)    
    imsave(f'{path}/{name}_DAPI.tif',DAPI_resize_array)   
    imsave(f'{path}/{name}_Ecad.tif',Ecad_resize_array)   
    imsave(f'{path}/{name}_CD45.tif',CD45_resize_array)   
    
    print(path)
    n=n+1

print('all done')
