==> 04_ mask tile merge 0325.ipynb <==
import glob
import math
from matplotlib import pyplot as plt
import napari
import numpy as np
from numpy import asarray

from skimage.io import imread,imshow,imsave
from skimage.filters import threshold_otsu
from skimage.measure import regionprops_table
import os
import pandas as pd
import PIL
from PIL import Image
import tifffile

from skimage.color import label2rgb
from skimage.segmentation import expand_labels,find_boundaries, mark_boundaries
os.chdir ('/Users/kim/Data/CRUK-OHSU_onedrive final 2022-09-30/Manuscript/IMC-IF method paper/Cell_report_method_resource/github')

print('library import done')

df=pd.read_csv('Analysis/new_July_2021/IFtilemask/IFtilelist.csv') # tile mask metadata 
df=df.sort_values('tileorder')
df['tileorder'][0]

slides=df['name'].drop_duplicates().tolist() # Whole slide image name 
slides

inputfolder='IFtilemask_mesmer'
preoutputfolder='maskmerge_mesmer_org'
outputfolder='maskmerge_mesmer_0325'

# Next 
# inputfolder='IFtilemask_mesmer_nuc'  
# preoutputfolder='maskmerge_mesmer_nuc_org' 
# outputfolder='maskmerge_mesmer_nuc_0325'  

mask_tile_list=sorted(glob.glob(f'Analysis/new_July_2021/{inputfolder}/*')) # mask tile location. 

for a in mask_tile_list:
    im=imread(a)
    print(a)
    print(im.shape)
    print(im.dtype)
print('done')

n=0
for slide in slides: #slide name 
    name=slide
    masks = []
    for lst in mask_tile_list: # slide name - mask tile sorting. 
        if name in lst:
            masks.append(lst)
    
    # ROI tile composition check 
    xnum=int(masks[-1].split('/')[-1].split('x_')[-1].split('_y_')[0])+1
    ynum=int(masks[-1].split('/')[-1].split('_y_')[-1].split('.tif')[0])+1

    # Mask array list   
    maskarrays=[np.array(imread(mask),np.uint32) for mask in masks]
    
    # Mask intensity renewal 
    prevmax = 0 
    inc_maskarrays=[]  
    for m in maskarrays:
        update= np.where(m>0,m+prevmax,0)
        inc_maskarrays.append(update)
        if m.any()> 0:
            prevmax=np.max(update)
        else:
            continue

    ### Mask intensity increase    
    start=0

    for array in inc_maskarrays:
        if np.max(array)>start:
            start=np.max(array)
    print("total cell number of ",name, "is", np.max(start))
    
    # Mask merge
    merge=np.vstack([np.hstack([inc_maskarrays[i*ynum+j] for j in range(ynum)]) for i in range(xnum)])
    tifffile.imsave(f'Analysis/new_July_2021/{preoutputfolder}/{name}.tiff',merge) # input merged mask folder location. 
    img = Image.fromarray(merge)

    # Mask resize 
    width=math.ceil(merge.shape[0]*0.325)
    height=math.ceil(merge.shape[1]*0.325)    
    imgresize=img.resize(size=(height,width),resample=PIL.Image.NEAREST)
    pix = np.array(imgresize)
    
    tifffile.imsave(f'Analysis/new_July_2021/{outputfolder}/{name}.tiff',pix) # input merged mask folder location. 
    image = imread(f'Analysis/new_July_2021/{outputfolder}/{name}.tiff') # for double-checking 
        
    print(name, f":{len(inc_maskarrays)} files loading done",'xnum:',xnum, 'ynum:',ynum)
    print('merged mask size:', merge.shape)
    print('image size after resizing:', pix.shape)
    print('cell number after resizing:', np.max(pix))
    print('image size after resizing:', pix.shape)
    print('image type after resizing:',image.dtype)
    print('')

    n=n+1
print('done')

### Merged File size Check

# PS14
IForgpath='Analysis/new_July_2021/CZItoTiff/IF/PS14.24197-1E.czi.tif'
maskmaergepath='Analysis/new_July_2021/maskmerge_mesmer_0325/PS14.24197-1E.tiff'
IMCregpath='Analysis/new_July_2021/IMC_Registration/registered_ROI001_023_PS14.24197-1E/Registered_191Ir_Ir191_DNA.ome.tiff'

viewer = napari.Viewer()

# IF original
image=imread(IForgpath)
viewer.add_image(image, name='IForiginal',blending='additive');
print("Original IF image sizse is:", image.shape)

# IFresize
image=imread(IForgpath)
width=math.ceil(image.shape[0]*0.325)
height=math.ceil(image.shape[1]*0.325)
img = Image.fromarray(image[:,:,0])
imgresize=img.resize(size=(height,width),resample=PIL.Image.NEAREST)
pix = np.array(imgresize)
viewer.add_image(pix, name='IFimgresize',blending='additive',colormap='green');
print("Resized IF image sizse is:", pix.shape)

# Maskmerge
maskimage=imread(maskmaergepath)
viewer.add_image(maskimage, name='Maskmerge',blending='additive',colormap='yellow');
print("Merged maskimage sizse is:", maskimage.shape)

nuc_marked = mark_boundaries(maskimage, maskimage, color=(1,0,1))
viewer.add_image(nuc_marked, name='Maskbound',blending='additive');

# IMCreg 
image=imread(IMCregpath)
viewer.add_image(image, name='IMCreg',blending='additive',colormap='red');
print("Registered IMC image sizse is:", image.shape)
