==> 05_Feature extraction_merge.ipynb <==
from glob import glob
from itertools import compress
import numpy as np
import os
import pandas as pd
from PIL import Image
from skimage.io import imread,imshow,imsave 
from skimage.measure import regionprops_table
import time
import warnings
warnings.filterwarnings(action='ignore')
Image.MAX_IMAGE_PIXELS = None

os.chdir ('/Users/kim/Data/CRUK-OHSU_onedrive final 2022-09-30/Manuscript/IMC-IF method paper/Cell_report_method_resource/github')
 
print('library import done')

### Feature extraction from cell mask (for mean intensity )


def feature_extraction(name,mask): 

    s = regionprops_table(mask, intensity_image=mask, 
               properties=('label','mean_intensity','area','centroid','eccentricity','axis_major_length','axis_minor_length' ))
     
    df = pd.DataFrame(s)
    cell_id = df['mean_intensity']
    area = df['area']
    centroid_X = df['centroid-1'] 
    centroid_Y = df['centroid-0']
    eccentricity = df['eccentricity']
    axis_major_length = df['axis_major_length']
    axis_minor_length = df['axis_minor_length']

    df = pd.concat([cell_id, area, centroid_X, centroid_Y, eccentricity, axis_major_length, axis_minor_length], axis=1, join='outer')
    df.rename(columns={'mean_intensity':'cell_id'}, inplace=True)
    df.rename(columns={'centroid-1':'centroid_X'}, inplace=True)
    df.rename(columns={'centroid-0':'centroid_Y'}, inplace=True)
    
    # IMC regionprop 
    img_fn = f'Analysis/new_July_2021/IMC_Registration/*{name}*/*ome.tiff'
    marker_list = sorted(glob(img_fn))[8:-4]  
    marker_name_list=[m.split('/')[-1].split('_')[-1].replace('.ome.tiff','') for m in marker_list] 
    
    i = 0        
    for i in range(len(marker_list)):
        marker=imread(marker_list[i])
        table = regionprops_table(mask, intensity_image = marker, properties=('label','mean_intensity','centroid'))
        df_table = pd.DataFrame(table)
        df_total = pd.concat([df,df_table],axis=1,join='outer')
        markername=marker_name_list[i]
        df_total.rename(columns={'mean_intensity': markername},inplace=True)
        df=df_total
        i=i+1 
        print(markername)
    print(name, 'done')
    return df
print('done')

#1. IF_cell
#2. IF_nuc_final
#3. IMC_cell 
#4. IMC_nuc_final 

### Feature extraction from nuclear mask (for nucleus area calculation)

def feature_extraction_nuc(name,mask): 

    s = regionprops_table(mask, intensity_image=mask, 
               properties=('label','mean_intensity','area','centroid','eccentricity' ))
     
    df = pd.DataFrame(s)
    cell_id = df['mean_intensity']
    area = df['area']
    centroid_X = df['centroid-1'] 
    centroid_Y = df['centroid-0']
    eccentricity = df['eccentricity']

    df = pd.concat([cell_id, area, centroid_X, centroid_Y, eccentricity], axis=1, join='outer')
    df.rename(columns={'mean_intensity':'cell_id'}, inplace=True)
    df.rename(columns={'centroid-1':'centroid_X'}, inplace=True)
    df.rename(columns={'centroid-0':'centroid_Y'}, inplace=True)
    df.rename(columns={'area':'area_nuc'}, inplace=True)

    return df
print('done')

def feature(name): 

    pt=name.split('.')[0]

    # IMC images 
    IMC_DNA = imread(glob(f'Analysis/new_July_2021/IMC_Registration/*{name}*/*191_DNA*')[0])
    IMC_Ecad = imread(glob(f'Analysis/new_July_2021/IMC_Registration/*{name}*/*E-Ca*')[0])

    # registered images    
    rCD45 = imread(glob(f'Data/IMC EMR (1)/*{name}*/resized_IF/*{name}*_CD45.tif')[0]) 
    rDAPI = imread(glob(f'Data/IMC EMR (1)/*{name}*/resized_IF/*{name}*_DAPI.tif')[0]) 
    rEcad = imread(glob(f'Data/IMC EMR (1)/*{name}*/resized_IF/*{name}*_Ecad.tif')[0]) 

    # masks 
    IF_cell = imread(f'Analysis/new_July_2021/maskmerge_mesmer_0325/{name}.tiff') 
    IF_nuc = imread(f'Analysis/new_July_2021/maskmerge_mesmer_nuc_0325/{name}.tiff')     
    IMC_cell = imread(f'Analysis/new_July_2021/IMC_mesmer/{pt}.tiff') 
    IMC_nuc = imread(f'Analysis/new_July_2021/IMC_mesmer_nuc/{pt}.tiff')

    IMC_rim = np.where(IMC_nuc != 0, 0, IMC_cell)
    IF_rim = np.where(IF_nuc != 0, 0, IF_cell)

    IF_nuc_final = np.where(IF_rim !=0, 0, IF_cell)
    IMC_nuc_final = np.where(IMC_rim !=0, 0, IMC_cell)

    # IF 
    IF_cell = feature_extraction (name= name, mask = IF_cell)
    IF_nuc= feature_extraction_nuc(name=name, mask= IF_nuc_final)

    img_fn = f'Analysis/new_July_2021/IMC_Registration/*{name}*/*ome.tiff'
    marker_list = sorted(glob(img_fn))[8:-4]  
    col1=[m.split('/')[-1].split('_')[-1].replace('.ome.tiff','') for m in marker_list] 
    col2 = ['cell_id', 'area', 'centroid_X','centroid_Y', 'eccentricity','axis_major_length','axis_minor_length']
    col=col2+col1
    IF_cell_final = IF_cell[col]

    IF_total= pd.merge(IF_cell_final, IF_nuc, on = ['cell_id'], how='inner')
    IF_total.rename(columns={'centroid_X_x':'centroid_X'}, inplace=True)
    IF_total.rename(columns={'centroid_Y_x':'centroid_Y'}, inplace=True)

    IF_total = IF_total.drop(columns = ['centroid_X_y','centroid_Y_y'])
    IF_total.rename(columns={'eccentricity_x':'eccentricity_cell'}, inplace=True)
    IF_total.rename(columns={'eccentricity_y':'eccentricity_nuc'}, inplace=True)
    IF_total.to_csv(f'Analysis/new_July_2021/ncratio_feature/IF_{name}.csv') 
    print('done')

    # IMC 
    IMC_cell = feature_extraction (name= name, mask = IMC_cell)
    IMC_nuc= feature_extraction_nuc(name=name, mask= IMC_nuc_final)

    img_fn = f'Analysis/new_July_2021/IMC_Registration/*{name}*/*ome.tiff'
    marker_list = sorted(glob(img_fn))[8:-4]  
    col1=[m.split('/')[-1].split('_')[-1].replace('.ome.tiff','') for m in marker_list] 
    col2 = ['cell_id', 'area', 'centroid_X','centroid_Y', 'eccentricity','axis_major_length','axis_minor_length']
    col=col2+col1
    IMC_cell_final = IMC_cell[col]

    IMC_total= pd.merge(IMC_cell_final, IMC_nuc, on = ['cell_id'], how='inner') # Matching cell and nuleus ID 
    IMC_total.rename(columns={'centroid_X_x':'centroid_X'}, inplace=True)
    IMC_total.rename(columns={'centroid_Y_x':'centroid_Y'}, inplace=True)

    IMC_total = IMC_total.drop(columns = ['centroid_X_y','centroid_Y_y'])
    IMC_total.rename(columns={'eccentricity_x':'eccentricity_cell'}, inplace=True)
    IMC_total.rename(columns={'eccentricity_y':'eccentricity_nuc'}, inplace=True)
    IMC_total.to_csv(f'Analysis/new_July_2021/ncratio_feature/IMC_{name}.csv') 
    
    print('done')
print('done')

%time
names = ['DE18.NE', 'DE28.NE.A', 'PS14.24197-1E', 'PS15.19650-B3', 'PS16.38634-F2']

for name in names:
    feature(name=name)
print('finally done')

### Feature table merge

# Feature merge IF

file_list= sorted(glob('Analysis/new_July_2021/ncratio_feature/IF_*'))

df_mask_column =pd.read_csv(file_list[0])
column_names=df_mask_column.columns
df_merg = pd.DataFrame(columns = column_names)

i=0
for i in range(len(file_list)):
    name= file_list[i].split('/')[-1].split('_')[-1].replace('.csv','')
    df_next = pd.read_csv(file_list[i])
    df_next['name']= name
    df_merg = pd.concat([df_merg,df_next], ignore_index=True)
    i=i+1   

df_merg.rename(columns={'area':'area_cell'}, inplace=True)
df_merg['area_rim'] = df_merg['area_cell'] - df_merg['area_nuc']
df_merg.to_csv('Analysis/new_July_2021/ncratio_feature/merge_IF.csv')
print('done')

# Feature merge IMC 

file_list= sorted(glob('Analysis/new_July_2021/ncratio_feature/IMC_*'))

df_mask_column =pd.read_csv(file_list[0])
column_names=df_mask_column.columns
df_merg = pd.DataFrame(columns = column_names)

i=0
for i in range(len(file_list)):
    name= file_list[i].split('/')[-1].split('_')[-1].replace('.csv','')
    df_next = pd.read_csv(file_list[i])
    df_next['name']= name
    df_merg = pd.concat([df_merg,df_next], ignore_index=True)
    i=i+1   

df_merg.rename(columns={'area':'area_cell'}, inplace=True)
df_merg['area_rim'] = df_merg['area_cell'] - df_merg['area_nuc']
df_merg.to_csv('Analysis/new_July_2021/ncratio_feature/merge_IMC.csv')
print('done')
